+++ scout/adapter/mongo/variant.py
@@ -4,7 +4,6 @@ import logging
 import re
 import pathlib
 import tempfile
-from scout.utils.requests import fetch_refseq_version
 
 from datetime import datetime
 from pprint import pprint as pp
@@ -140,7 +139,7 @@ class VariantHandler(VariantLoader):
                 if not hgnc_transcript.get('refseq_id'):
                     continue
 
-                refseq_id = fetch_refseq_version(hgnc_transcript['refseq_id'])
+                refseq_id = hgnc_transcript['refseq_id']
                 transcript['refseq_id'] = refseq_id
                 variant_obj['has_refseq'] = True
                 # Check if the refseq id are disease associated