+++ CHANGELOG.md
@@ -13,12 +13,14 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Search for cases phenotypically similar to a case also from WUI.
 - Case variant searches can be limited to similar cases, matching HPO-terms,
   phenogroups and cohorts.
-- Dearchive reruns and flag them as 'inactive' if archived
+- De-archive reruns and flag them as 'inactive' if archived
 - Sort cases by analysis_date, track or status
 - Display cases in the following order: prioritized, active, inactive, archived, solved
 - Assign case to user when user activates it or asks for rerun
 - Case becomes inactive when it has no assignees
-- Fetch refseq version from entrez and use it in clinvar formr
+- Fetch refseq version from entrez and use it in clinvar form
+- Showing SV variant annotations: SV cgh frequencies, gnomad-SV, local SV frequencies
+- Showing transcripts mapping score in segmental duplications
 
 ### Fixed
 - Fixes of and induced by build tests